improve_the_research_feature_by_adding_like_not_search_by_the_exact_string

+++ resources/views/apppages/projects/oneproject.blade.php
@@ -52,28 +52,27 @@
         </form>
         <div class="ml-auto py-12 ">
             <div class="max-w-7xl mx-auto sm:px-6 lg:px-8">
-
                 <div class="grid grid-cols-1 md:grid-cols-1 lg:grid-cols-3 gap-6">
-
                     @forelse ($foundprojects as $project)
-                        <div id="project{{ $project->id }}"  style="outline-style: solid;
-                            outline-color: {{$project->responsibility->color}};"
-                            class="px-10 bg-white overflow-hidden shadow-xl sm:rounded-lg min-w-full h-40 flex flex-col items-start justify-around">
-                            <a href="{{ route('showproject', $project->id) }}">
-                                <p class="">{{ $project->name }} </p>
-                            </a>
-                            <div class="flex flex-row gap-1">
-                                <a href="{{ route('editproject', $project->id) }}"
-                                    class="w-6 h-6 hover:w-8 hover:h-8"> <img src="{{ asset('images/edit.png') }}"
-                                        alt="editbutton">
+                        @if ($project->responsibility->user_id === $authid)
+                            <div id="project{{ $project->id }}" style="outline-style: solid;
+                            outline-color: {{ $project->responsibility->color }};"
+                                class="px-10 bg-white overflow-hidden shadow-xl sm:rounded-lg min-w-full h-40 flex flex-col items-start justify-around">
+                                <a href="{{ route('showproject', $project->id) }}">
+                                    <p class="">{{ $project->name }} </p>
                                 </a>
-                                <a href="javascript:void(0)" onclick="deleteProject({{ $project->id }})"
-                                    class="w-6 h-6 hover:w-8 hover:h-8"> <img src="{{ asset('images/delete.png') }}"
-                                        alt="deltebutton">
-                                </a>
+                                <div class="flex flex-row gap-1">
+                                    <a href="{{ route('editproject', $project->id) }}"
+                                        class="w-6 h-6 hover:w-8 hover:h-8"> <img src="{{ asset('images/edit.png') }}"
+                                            alt="editbutton">
+                                    </a>
+                                    <a href="javascript:void(0)" onclick="deleteProject({{ $project->id }})"
+                                        class="w-6 h-6 hover:w-8 hover:h-8"> <img
+                                            src="{{ asset('images/delete.png') }}" alt="deltebutton">
+                                    </a>
+                                </div>
                             </div>
-
-                        </div>
+                        @endif
 
                     @empty
                         <div
